# docs/api/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: GHARMIND AI — Household Operating System API
  description: |
    API for GHARMIND AI — India's First AI-powered Household Operating System.
    
    Powers the Digital Twin, Prediction Engine, What-If Simulator, and 
    conversational household AI (Gharji).
    
    All timestamps are in ISO 8601 format with Asia/Kolkata (IST) timezone.
    Authentication via AWS Cognito JWT.
  version: "1.0.0"
  contact:
    name: GHARMIND AI Team
    email: dev@gharmind.in
  license:
    name: Proprietary
servers:
  - url: https://api.gharmind.in/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Development
security:
  - CognitoJWT: []
...
paths:
...
  /households:
    post:
      tags: [Households]
      summary: Create a new household
      operationId: createHousehold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseholdRequest'
      responses:
        '201':
          description: Household created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '422':
          $ref: '#/components/responses/ValidationError'
  /households/{householdId}:
    get:
      tags: [Households]
      summary: Get household details
      operationId: getHousehold
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
      responses:
        '200':
          description: Household details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Households]
      summary: Update household settings
      operationId: updateHousehold
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHouseholdRequest'
      responses:
        '200':
          description: Household updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
...
  /households/{householdId}/predictions:
    get:
      tags: [Predictions]
      summary: Get current prediction feed
      operationId: getPredictions
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
        - name: horizon
          in: query
          schema:
            type: string
            enum: [30min, 2h, today, week]
            default: 2h
        - name: priority
          in: query
          schema:
            type: string
            enum: [critical, high, all]
            default: all
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Prediction feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionFeed'
  /households/{householdId}/predictions/{predictionId}:
    get:
      tags: [Predictions]
      summary: Get specific prediction with full reasoning
      operationId: getPrediction
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
        - name: predictionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Prediction detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionDetail'
  /households/{householdId}/predictions/{predictionId}/feedback:
    post:
      tags: [Predictions]
      summary: Rate a prediction (feedback loop)
      operationId: ratePrediction
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
        - name: predictionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionFeedbackRequest'
      responses:
        '200':
          description: Feedback recorded
  /households/{householdId}/predictions/refresh:
    post:
      tags: [Predictions]
      summary: Force regenerate predictions now
      operationId: refreshPredictions
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
      responses:
        '200':
          description: Predictions refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionFeed'

  # ─── SIMULATOR ───────────────────────────────────────────────
...
  /households/{householdId}/routines:
    get:
      tags: [Routines]
      summary: List all household routines
      operationId: listRoutines
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
        - name: type
          in: query
          schema:
            type: string
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of routines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineList'

    post:
      tags: [Routines]
      summary: Create a new routine
      operationId: createRoutine
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoutineRequest'
      responses:
        '201':
          description: Routine created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Routine'
  /households/{householdId}/routines/upcoming:
    get:
      tags: [Routines]
      summary: Get routines in next 24 hours
      operationId: getUpcomingRoutines
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
            maximum: 168
      responses:
        '200':
          description: Upcoming routines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpcomingRoutinesResponse'

  # ─── INSIGHTS ────────────────────────────────────────────────
...
components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito Access Token
  parameters:
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Household UUID
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

    # ── Household ─────────────────────────────────────────────
    Household:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Sharma Residence"
        city:
          type: string
          example: "Pune"
        state:
          type: string
          example: "Maharashtra"
        home_type:
          type: string
          enum: [apartment, villa, row_house, independent]
        language_preference:
          type: string
          enum: [en, hi, hinglish]
          default: hinglish
        ai_persona_name:
          type: string
          default: "Gharji"
        twin_initialized:
          type: boolean
        subscription_tier:
          type: string
          enum: [free, pro, family]
        created_at:
          type: string
          format: date-time
    CreateHouseholdRequest:
      type: object
      required: [name, city, state]
      properties:
        name:
          type: string
          maxLength: 200
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        home_type:
          type: string
          enum: [apartment, villa, row_house, independent]
        language_preference:
          type: string
          enum: [en, hi, hinglish]
          default: hinglish
    UpdateHouseholdRequest:
      type: object
      properties:
        name:
          type: string
        language_preference:
          type: string
          enum: [en, hi, hinglish]
        ai_persona_name:
          type: string
...
    PredictionFeed:
      type: object
      properties:
        household_id:
          type: string
          format: uuid
        generated_at:
          type: string
          format: date-time
        predictions:
          type: array
          items:
            $ref: '#/components/schemas/Prediction'
        summary:
          type: object
          properties:
            critical_count:
              type: integer
            high_count:
              type: integer
            normal_count:
              type: integer
    Prediction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [routine_start, appliance_action, member_arrival, festival_prep, power_event, water_event, meal_time, study_reminder]
        title:
          type: string
        description:
          type: string
        action_suggestion:
          type: string
          description: Actionable recommendation in user's language
        predicted_for:
          type: string
          format: date-time
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        priority:
          type: string
          enum: [critical, high, normal, low]
        category:
          type: string
        expires_at:
          type: string
          format: date-time
    PredictionDetail:
      allOf:
        - $ref: '#/components/schemas/Prediction'
        - type: object
          properties:
            reasoning:
              type: string
              description: Claude's reasoning chain
            evidence:
              type: array
              items:
                type: string
              description: Supporting data points
    PredictionFeedbackRequest:
      type: object
      required: [feedback]
      properties:
        feedback:
          type: string
          enum: [helpful, not_helpful, wrong]
        note:
          type: string
          maxLength: 500

    # ── Simulator ─────────────────────────────────────────────
...
    Routine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        routine_type:
          type: string
        recurrence:
          type: string
        is_active:
          type: boolean
        is_ai_detected:
          type: boolean
        confidence_score:
          type: number
        next_expected_at:
          type: string
          format: date-time
        last_executed_at:
          type: string
          format: date-time
    RoutineList:
      type: object
      properties:
        total:
          type: integer
        routines:
          type: array
          items:
            $ref: '#/components/schemas/Routine'
    CreateRoutineRequest:
      type: object
      required: [name, routine_type, recurrence, schedule_expression]
      properties:
        name:
          type: string
          maxLength: 200
        routine_type:
          type: string
          enum: [pooja, motor, tuition, meal, study, chai, cleaning, exercise, commute, festival]
        recurrence:
          type: string
          enum: [daily, weekly, monthly, annual, conditional]
        schedule_expression:
          type: object
        primary_member_id:
          type: string
          format: uuid
    UpcomingRoutinesResponse:
      type: object
      properties:
        routines:
          type: array
          items:
            type: object
            properties:
              routine:
                $ref: '#/components/schemas/Routine'
              expected_at:
                type: string
                format: date-time
              confidence:
                type: number

    # ── Insights ──────────────────────────────────────────────
...